moved the render to a separate module
